LPD-59495 Move info_panel to main_view
