+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/info_panel/context/index.ts
@@ -8,7 +8,7 @@ import React from 'react';
 import {
 	IAssetInformation,
 	IAssetObjectEntry,
-} from '../../../../structure_builder/types/AssetType';
+} from '../../../structure_builder/types/AssetType';
 
 export interface IAssetTypeInfoPanelContext extends IAssetInformation {
 	objectEntries?: IAssetObjectEntry[];
