+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/info_panel/tab_content/VersionsTabContent.tsx
