+++ modules/apps/site/site-cms-site-initializer/test/js/main_view/info_panel/mocks/index.ts
